+++ generators/data-module/templates/scripts/_Model.js
@@ -1,0 +1,3 @@
+angular
+  .module('<%= moduleName %>')
+  .constant('<%= _.capitalize(resourceName) %>', supersonic.data.model('<%= _.capitalize(resourceName) %>'));