+++ generators/module/templates/scripts/_IndexController.js
@@ -1,0 +1,5 @@
+angular
+  .module('{{ moduleName }}')
+  .controller('IndexController', function($scope, supersonic) {
+    // Controller functionality here
+  });